+++ wxc_sdk/__init__.py
@@ -33,7 +33,7 @@ from dataclasses import dataclass
 
 __all__ = ['WebexSimpleApi']
 
-__version__ = '1.11.0'
+__version__ = '1.12.0'
 
 log = logging.getLogger(__name__)