+++ docs/changes.rst
@@ -1,10 +1,13 @@
 Release history
 ===============
 
+1.12.0
+------
 - feat: new attribute call_park_extension in :class:`wxc_sdk.telephony.callpark.CallPark`
 - feat: new parameters details, restricted_non_geo_numbers for :meth:`wxc_sdk.telephony.TelephonyApi.phone_numbers`
 - feat: new Api :class:`wxc_sdk.telephony.location.receptionist_contacts.ReceptionistContactsDirectoryApi`
-- fix: correct support for enum URL params in :method:`wxc_sdk.workspaces.WorkspacesApi.list`
+- fix: correct support for enum URL params in :meth:`wxc_sdk.workspaces.WorkspacesApi.list`
+- feat: new attribute :attr:`wxc_sdk.telephony.autoattendant.AutoAttendantMenu.audio_file`
 
 1.11.0
 ------